LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollIntoView-scrolling-container-expected.txt


FAIL scrollIntoView on a scrolling container should scroll its outer scrollers into view but not scroll itself. assert_equals: Should scroll inner scroller to scroll margin offset. expected 20 but got 0
